Save a np.uint8 image to test tif compression
The current default for random images is np.uint64, which causes
problems with conversion to float.

+++ tests/test_io.py
@@ -3,7 +3,7 @@ from tempfile import NamedTemporaryFile
 from microscopium import io as mio
 
 def test_imsave_tif_compress():
-    im = np.random.randint(0, 256, size=(1024, 1024, 3))
+    im = np.random.randint(0, 256, size=(1024, 1024, 3)).astype(np.uint8)
     with NamedTemporaryFile(suffix='.tif') as fout:
         fname = fout.name
         fout.close()
